requirements.txt
# requirements.txt
Django==3.2
psycopg2-binary==2.9.3
djangorestframework==3.12.4
pytest-django==4.4.0
requests==2.27.1
redis==3.5.3
django-redis==5.0.0
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
